refactor(svelte2tsx): svelte2tsx() の E2E テストを結合テストへ切り出し
`mod tests` のうち公開 API だけを使う E2E テスト 30 本を
`crates/rsvelte_core/tests/svelte2tsx_entry.rs` へ移す。
ヘルパー単体テストは、対応するヘルパーと同じコミットで移送先へ移す。

純粋な移動でロジック変更はなく、出力はバイト単位で不変。
